Deleted git changelist support
This was used to limit the compile_commands to specific git changelists. While something like this may be needed in the future this will need to be rewritten for perforce. The --commit arguments still exist but I have commented out the implementation for now until we know what is going to be needed for this fully. The main reason to remove this is  to remove the dependencies and allow us to clean up our python packages